// LR_4/my-rent-service/src/types/offer.ts
type OfferLocation = {
    latitude: number;
    longitude: number;
    zoom: number;
};

type CityOffer = {
    name: string;
    location: OfferLocation;
};

type HostOffer = {
    name: string;
    avatarUrl: string;
    isPro: boolean;
};

export type FullOffer = {
    id: string;
    title: string;
    type: string;
    price: number;
    city: CityOffer;
    location: OfferLocation;
    isFavorite: boolean;
    isPremium: boolean;
    rating: number;
    description: string;
    bedrooms: number;
    goods: string[];
    host: HostOffer;
    images: string[];
    maxAdults: number;
};

export type OffersList = {
    isPremium: any;
    id: string;
    title: string;
    type: string;
    price: number;
    city: CityOffer;
    location:OfferLocation;
    isFavorite: boolean;
    rating: number;
    previewImage: string; 
  };